requirements.txt
numpy>=1.18.2
pillow>=7.1.0
pydaedalus>=0.1.5
matplotlib>=3.2.1
celluloid>=0.2.0
click>=7.1.2
